refactor: ReleaseTemplatesFeedback component and integrate it into FeatureOverviewEnvironment
