+++ frontend/src/component/feature/FeatureStrategy/FeatureStrategyMenu/FeatureStrategyMenu.tsx
@@ -5,7 +5,7 @@ import PermissionButton, {
     type IPermissionButtonProps,
 } from 'component/common/PermissionButton/PermissionButton';
 import { CREATE_FEATURE_STRATEGY } from 'component/providers/AccessProvider/permissions';
-import { Popover, styled, Link } from '@mui/material';
+import { Popover, styled } from '@mui/material';
 import { FeatureStrategyMenuCards } from './FeatureStrategyMenuCards/FeatureStrategyMenuCards';
 import { formatCreateStrategyPath } from '../FeatureStrategyCreate/FeatureStrategyCreate';
 import MoreVert from '@mui/icons-material/MoreVert';
@@ -20,9 +20,7 @@ import { useReleasePlans } from 'hooks/api/getters/useReleasePlans/useReleasePla
 import { useChangeRequestsEnabled } from 'hooks/useChangeRequestsEnabled';
 import { formatUnknownError } from 'utils/formatUnknownError';
 import { useUiFlag } from 'hooks/useUiFlag';
-import type { Link as RouterLink } from 'react-router-dom';
 import useUiConfig from 'hooks/api/getters/useUiConfig/useUiConfig';
-import { RELEASE_TEMPLATE_FEEDBACK } from 'constants/links';
 
 interface IFeatureStrategyMenuProps {
     label: string;
@@ -37,10 +35,8 @@ interface IFeatureStrategyMenuProps {
 }
 
 const StyledStrategyMenu = styled('div')(({ theme }) => ({
-    flexShrink: 0,
     display: 'flex',
     flexFlow: 'row',
-    flex: 1,
     justifyContent: 'flex-end',
     gap: theme.spacing(1),
 }));
@@ -50,15 +46,6 @@ const StyledAdditionalMenuButton = styled(PermissionButton)(({ theme }) => ({
     width: theme.spacing(4.5),
     alignSelf: 'stretch',
     paddingBlock: 0,
-}));
-
-const StyledLink = styled(Link<typeof RouterLink | 'a'>)(({ theme }) => ({
-    display: 'flex',
-    alignItems: 'center',
-    gap: theme.spacing(1),
-    color: theme.palette.links,
-    fontWeight: theme.typography.fontWeightMedium,
-    textDecoration: 'none',
 }));
 
 const Spacer = styled('div')(({ theme }) => ({
@@ -178,40 +165,27 @@ export const FeatureStrategyMenu = ({
     );
 
     return (
-        <StyledStrategyMenu onClick={(event) => event.stopPropagation()}>
+        <StyledStrategyMenu
+            onClick={(event) => event.stopPropagation()}
+            data-testid='????'
+        >
             {displayReleasePlanButton ? (
-                <>
-                    {allowReleasePlanFeedback ? (
-                        <>
-                            <StyledLink
-                                component='a'
-                                href={RELEASE_TEMPLATE_FEEDBACK}
-                                underline='hover'
-                                rel='noopener noreferrer'
-                                target='_blank'
-                            >
-                                Give feedback to release templates
-                            </StyledLink>
-                            <Spacer />
-                        </>
-                    ) : null}
-                    <PermissionButton
-                        data-testid='ADD_TEMPLATE_BUTTON'
-                        permission={CREATE_FEATURE_STRATEGY}
-                        projectId={projectId}
-                        environmentId={environmentId}
-                        onClick={openReleasePlans}
-                        aria-labelledby={popoverId}
-                        variant='outlined'
-                        sx={{ minWidth: matchWidth ? '282px' : 'auto' }}
-                        disabled={Boolean(disableReason)}
-                        tooltipProps={{
-                            title: disableReason ? disableReason : undefined,
-                        }}
-                    >
-                        Use template
-                    </PermissionButton>
-                </>
+                <PermissionButton
+                    data-testid='ADD_TEMPLATE_BUTTON'
+                    permission={CREATE_FEATURE_STRATEGY}
+                    projectId={projectId}
+                    environmentId={environmentId}
+                    onClick={openReleasePlans}
+                    aria-labelledby={popoverId}
+                    variant='outlined'
+                    sx={{ minWidth: matchWidth ? '282px' : 'auto' }}
+                    disabled={Boolean(disableReason)}
+                    tooltipProps={{
+                        title: disableReason ? disableReason : undefined,
+                    }}
+                >
+                    Use template
+                </PermissionButton>
             ) : null}
 
             <PermissionButton
